requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
attrs==22.1.0
build==0.10.0
CacheControl
cachy
certifi
cffi
cfgv
chardet
charset-normalizer
cleo
click
clikit
cmake
colorama
crashtest
crispy-bootstrap5
cryptography
cssselect2
defusedxml
diff-match-patch==20200713
distlib==0.3.5
dj-database-url==1.0.0
Django==4.1.1
django-allauth==0.51.0
django-appconf==1.0.5
django-autocomplete-light==3.9.4
django-compressor==4.1
django-cors-headers==3.13.0
django-countries==7.4.2
django-crispy-forms==1.14.0
django-environ
django-filter==22.1
django-grappelli==3.0.3
django-import-export==2.9.0
django-js-asset==2.0.0
django-libsass==0.9
django-mptt==0.14.0
django-test-plus==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
docopt==0.6.2
drf-nested-field-multipart==0.0.3
drf-nested-routers==0.93.4
drf-spectacular==0.24.2
drf-writable-nested==0.7.0
dulwich==0.21.3
setuptools
python-dotenv
et-xmlfile==1.1.0
filelock==3.8.0
future==0.18.2
html5lib==1.1
identify==2.5.17
idna==3.4
imageio==2.24.0
importlib-metadata==5.0.0
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
installer==0.7.0
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.16.0
keyring==23.13.1
libsass==0.21.0
lit==16.0.2
lockfile==0.12.2
lxml==4.9.1
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.4
mypy-extensions==0.4.3
networkx==3.0
nodeenv==1.7.0
num2words==0.5.12
numpy==1.23.5
oauthlib==3.2.1
odfpy==1.4.1
opencv-python-headless==4.5.4.60
openpyxl==3.0.10
oscrypto==1.3.0
packaging==23.0
pandas
pastel
pexpect
Pillow
pkginfo
platformdirs
poetry
poetry-core
poetry-plugin-export
pre-commit
psycopg2
ptyprocess==0.7.0
pycparser==2.21
pyHanko==0.14.0
pyhanko-certvalidator==0.19.5
PyJWT==2.5.0
pylev==1.4.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.18.1
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz
pytz-deprecation-shim==0.1.0.post0
PyWavelets
pywin32-ctypes
PyYAML
qrcode
rapidfuzz
rcssmin
reportlab
requests
requests-oauthlib
requests-toolbelt
rjsmin==1.2.0
scikit-image
scipy
SecretStorage
shellingham
six
sqlparse
stripe
svglib
sympy
tablib
tifffile
tinycss2
tokenize-rt
toml
tomli
tomlkit
tqdm
trove-classifiers
types-cryptography
typing_extensions
tzdata
tzlocal
uritemplate
uritools
urllib3
virtualenv
Wand
webencodings
xhtml2pdf
xlrd
XlsxWriter
xlwt
zipp
